de-dupe createPublicationData
